Move xk6-dashboard code into internal/dashboard
Copy the dashboard package from github.com/grafana/xk6-dashboard into
internal/dashboard/ so the code lives directly in this repository.
Update internal/cmd/outputs.go and internal/cmd/ui.go to import from
the new internal path. Remove xk6-dashboard as an external dependency;
xk6-dashboard-assets remains external as it holds the UI assets in a
separate repository.
