+++ internal/dashboard/assets.go
