+++ internal/dashboard/registry.go
